Add shared XRHelpers move some shared code there, and address a few more comments.
